+++ src/PuppeteerSharp.Contrib.Extensions.Unsafe/PuppeteerSharp.Contrib.Extensions.Unsafe.csproj
@@ -2,6 +2,11 @@
 
   <PropertyGroup>
     <TargetFramework>netstandard2.0</TargetFramework>
+    <RootNamespace>PuppeteerSharp.Contrib.Extensions</RootNamespace>
   </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\PuppeteerSharp.Contrib.Extensions\PuppeteerSharp.Contrib.Extensions.csproj" />
+  </ItemGroup>
 
 </Project>
