+++ src/PuppeteerSharp.Contrib.Should.Unsafe/InternalExtensions.cs
@@ -1,0 +1,12 @@
+using System.Threading.Tasks;
+
+namespace PuppeteerSharp.Contrib.Should
+{
+    internal static class InternalExtensions
+    {
+        internal static T Result<T>(this Task<T> task)
+        {
+            return task.GetAwaiter().GetResult();
+        }
+    }
+}
