+++ src/PuppeteerSharp.Contrib.Should.Unsafe/PuppeteerSharp.Contrib.Should.Unsafe.csproj
@@ -2,6 +2,12 @@
 
   <PropertyGroup>
     <TargetFramework>netstandard2.0</TargetFramework>
+    <RootNamespace>PuppeteerSharp.Contrib.Should</RootNamespace>
   </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\PuppeteerSharp.Contrib.Extensions.Unsafe\PuppeteerSharp.Contrib.Extensions.Unsafe.csproj" />
+    <ProjectReference Include="..\PuppeteerSharp.Contrib.Should\PuppeteerSharp.Contrib.Should.csproj" />
+  </ItemGroup>
 
 </Project>
